created useSubcribeConversation hook, extracted from Conversation component
